refactor: simplify di, normalize strings when filtering
