+++ data/src/main/java/com/jcaique/dialetus/data/mappers/RegionsMapper.kt
@@ -4,6 +4,7 @@ import com.jcaique.dialetus.data.responses.RegionResponse
 import com.jcaique.dialetus.domain.models.Region
 
 internal object RegionsMapper {
+
     fun toDomain(response: RegionResponse): Region =
         Region(response.name, response.total)
 }
